+++ client-app/src/components/postDetails/PostDetails.jsx
@@ -1,37 +1,67 @@
 import { useEffect, useState, useContext } from "react"
 import { useParams, useNavigate } from "react-router-dom"
 
-import { getPostsDetails, deletePost } from "../../services/postFirestoreService"
 import AuthContext from "../../contexts/authContext";
+import ErrorContext from "../../contexts/errorContext"
+
+import { getPostsDetails } from "../../services/getPostService"
+import { deletePost } from "../../services/otherPostServices";
+import { addToSaved, checkForSaved, removeFromSaved } from "../../services/savedService";
 
 import Comments from "../comments/Comments";
-import AddToFavourites from "../addToFavourites/AddToFavourites";
-import AddComment from "../addComment/AddComment";
 
 function PostDetails() {
-    const [dataState, setDataState] = useState({})
-    const { userId } = useContext(AuthContext)
-    const navigate = useNavigate()
+    const { showErrorHandler } = useContext(ErrorContext)
+    const { userId, fullName, isAuthenticated } = useContext(AuthContext)
 
-    const { isAuthenticated } = useContext(AuthContext)
+    const [dataState, setDataState] = useState({})
+    const [showComments, setShowComments] = useState(false)
+    const [saveState, setSaveState] = useState({ canBeSaved: false, canBeUnSaved: false })
+
+    const navigate = useNavigate()
 
 
     const { postId } = useParams()
 
     function deletePostHandler() {
+
+        if (!window.confirm("Are you sure you want to delete this post?")) return
         deletePost(postId)
             .then(() => {
                 navigate('/home')
             })
             .catch(error => {
                 console.error("Error while deleting post at PostDetails.jsx: ", error)
+                showErrorHandler('Error while deleting post!')
             })
     }
 
     function redirectToEdit() {
+        if (!window.confirm("Are you sure you want to edit this post?")) return
         navigate(`/posts/edit/${postId}`)
     }
 
+    function addSavedHandler() {
+        addToSaved({ dataState, userId }).then(result => {
+            if (result) {
+                setSaveState({ canBeSaved: false, canBeUnSaved: true })
+            }
+        }).catch(error => {
+            showErrorHandler('Error while saving post!')
+            console.error("Error while saving post at PostDetails.jsx: ", error)
+        })
+    }
+
+    function removeSavedHandler() {
+        removeFromSaved({ postId, userId }).then(result => {
+            if (result) {
+                setSaveState({ canBeSaved: true, canBeUnSaved: false })
+            }
+        }).catch(error => {
+            showErrorHandler('Error while unsaving post!')
+            console.error("Error while unsaving post at PostDetails.jsx: ", error)
+        })
+    }
 
     const formatedDate = (date) => {
         if (date instanceof Date) {
@@ -43,6 +73,11 @@ function PostDetails() {
         }
     }
 
+    const showCommentsHandler = () => {
+        let oldState = showComments
+        setShowComments(!oldState)
+    }
+
     useEffect(() => {
         getPostsDetails(postId)
             .then(result => {
@@ -50,8 +85,21 @@ function PostDetails() {
             })
             .catch(error => {
                 console.error("Error while getting post details at PostDetails.jsx: ", error)
+                showErrorHandler('Error while getting post details!')
             })
     }, [postId, navigate])
+
+
+    useEffect(() => {
+        if (!userId) return
+        checkForSaved({ postId, userId })
+            .then(({ canBeSaved, canBeUnSaved }) => {
+                setSaveState({ canBeSaved, canBeUnSaved })
+            })
+            .catch(error => {
+                console.error("Error while checking for saved posts: ", error)
+            })
+    }, [userId, postId])
 
     return (
         <div className="w-full md:w-2/5 mx-auto">
@@ -67,34 +115,10 @@ function PostDetails() {
 
             </div>
 
-            {/* <div className="mx-5">
-                <img
-                    src="https://static.politico.com/dims4/default/fcd6d6a/2147483647/resize/1920x/quality/90/?url=https%3A%2F%2Fstatic.politico.com%2F22%2F87%2F2259ffd444678054896b9fa32b4d%2Fgettyimages-1221513169.jpg"
-                    alt="Georgia Gov. Brian Kemp speaks to the media during a press conference. | Kevin C. Cox/Getty Images" />
-            </div> */}
-
-            {/* <div className="w-full text-gray-600 text-normal mx-5">
-                <p className="border-b py-3">Georgia Gov. Brian Kemp speaks to the media during a press conference. | Kevin C.
-                    Cox/Getty Images</p>
-            </div> */}
-
             <div className="w-full text-gray-600 font-thin italic px-5 pt-3">
-                By <strong className="text-gray-700">Posted by user name {dataState.publisherName}</strong><br />
+                By <strong className="text-gray-700">Posted by {dataState.creatorName}</strong><br />
 
                 {formatedDate(dataState.createdAt)}<br />
-                {/* Updated: 07/17/2020 10:33 AM EDT */}
-
-                <span class="isolate inline-flex rounded-md shadow-sm">
-                    {/* <button type="button" class="relative inline-flex items-center rounded-l-md bg-white px-3 py-2 text-sm font-semibold text-gray-900 ring-1 ring-inset ring-gray-300 hover:bg-gray-50 focus:z-10">share</button> */}
-                    {dataState.ownerId === userId && (
-                        <>
-                            <button type="button" onClick={redirectToEdit} class="relative -ml-px inline-flex items-center bg-white px-3 py-2 text-sm font-semibold text-gray-900 ring-1 ring-inset ring-gray-300 hover:bg-gray-50 focus:z-10">Edit</button>
-                            <button type="button" monClick={deletePostHandler} class="relative -ml-px inline-flex items-center rounded-r-md bg-white px-3 py-2 text-sm font-semibold text-gray-900 ring-1 ring-inset ring-gray-300 hover:bg-gray-50 focus:z-10">Delete</button>
-                        </>
-                    )}
-                </span>
-
-
 
             </div>
 
@@ -103,8 +127,65 @@ function PostDetails() {
 
             </div>
 
-            <div>
-                <Comments currentUser={userId} />
+            <span className="isolate inline-flex rounded-md shadow-sm">
+                <>
+                    <div id="crudButtons"
+                        className="flex flex-col items-center justify-center">
+
+                        <div className="flex h-20 items-center justify-center gap-1 pb-3">
+
+                            <button
+                                onClick={showCommentsHandler}
+                                className="inline-flex items-center gap-2 rounded-full border border-[#7629c8] px-6 py-2 text-sm font-semibold text-[#7629c8] transition-all hover:bg-[#7629c8] hover:text-white hover:shadow-lg active:scale-95 disabled:pointer-events-none disabled:opacity-50 disabled:shadow-none"
+                            >
+                                {showComments ? 'Hide comments' : 'Show comments'}
+                            </button>
+
+
+                            {dataState.ownerId === userId && (
+                                <>
+                                    <button onClick={redirectToEdit}
+                                        className="inline-flex items-center gap-2 rounded-full border border-[#7629c8] px-6 py-2 text-sm font-semibold text-[#7629c8] transition-all hover:bg-[#7629c8] hover:text-white hover:shadow-lg active:scale-95 disabled:pointer-events-none disabled:opacity-50 disabled:shadow-none"
+                                    >
+                                        Edit post
+                                    </button>
+
+                                    <button
+                                        onClick={deletePostHandler}
+                                        className="inline-flex items-center gap-2 rounded-full border border-[#7629c8] px-6 py-2 text-sm font-semibold text-[#7629c8] transition-all hover:bg-[#7629c8] hover:text-white hover:shadow-lg active:scale-95 disabled:pointer-events-none disabled:opacity-50 disabled:shadow-none"
+                                    >
+                                        Delete post
+                                    </button>
+                                </>
+                            )}
+
+                            {isAuthenticated && saveState.canBeSaved &&
+                                <button
+                                    onClick={addSavedHandler}
+                                    className="inline-flex items-center gap-2 rounded-full border border-[#7629c8] px-6 py-2 text-sm font-semibold text-[#7629c8] transition-all hover:bg-[#7629c8] hover:text-white hover:shadow-lg active:scale-95 disabled:pointer-events-none disabled:opacity-50 disabled:shadow-none"
+                                >
+                                    Save post
+                                </button>
+                            }
+
+                            {isAuthenticated && saveState.canBeUnSaved &&
+                                <button
+                                    onClick={removeSavedHandler}
+                                    className="inline-flex items-center gap-2 rounded-full border border-[#7629c8] px-6 py-2 text-sm font-semibold text-[#7629c8] transition-all hover:bg-[#7629c8] hover:text-white hover:shadow-lg active:scale-95 disabled:pointer-events-none disabled:opacity-50 disabled:shadow-none"
+                                >
+                                    Unsave post
+                                </button>
+                            }
+                        </div>
+
+
+                    </div>
+                </>
+
+            </span>
+
+            <div className={`pt-6 ` + (showComments ? 'block' : 'hidden')}>
+                <Comments currentId={userId} creatorName={fullName} isAuthenticated={isAuthenticated} />
             </div>
 
 
@@ -115,53 +196,3 @@ function PostDetails() {
 }
 
 export default PostDetails
-
-
-
-
-
-function oldCode() {
-    return (
-        <article classNameName="blog-post px-3 py-5 p-md-5">
-            <div classNameName="container">
-
-                <header classNameName="blog-post-header">
-                    <h2 classNameName="title mb-2">{dataState.title}</h2>
-                    <div classNameName="meta mb-3">
-                        <span classNameName="date">Published on {formatedDate(dataState.createdAt)}</span>
-
-                        {dataState.ownerId === userId && (
-                            <>
-                                <button onClick={redirectToEdit} style={{ color: 'white', backgroundColor: 'red', border: 'none', padding: '10px', cursor: 'pointer' }}>Edit</button>{' '}
-                                <button onClick={deletePostHandler} style={{ color: 'white', backgroundColor: 'red', border: 'none', padding: '10px', cursor: 'pointer' }}>Delete</button>{' '}
-                            </>
-                        )}
-
-                    </div>
-                </header>
-
-                <div classNameName="blog-post-body">
-                    <figure classNameName="blog-banner">
-
-                    </figure>
-                    <p>
-                        {dataState.body}
-                    </p>
-                </div>
-
-                {dataState.id && <AddToFavourites
-                    userId={userId}
-                    postId={dataState.id}
-                    ownerId={dataState.ownerId}
-                    title={dataState.title}
-                    body={dataState.body}
-                />}
-
-                <div classNameName="blog-comments-section">
-                    <Comments currentUser={userId} />
-                </div>
-
-            </div>
-        </article>
-    )
-}
